chore: created directory, separated types, constants, interfaces, resolved other feedback
